refactor(#7): изменены пакеты в модуле common
