+++ .idea/misc.xml
@@ -12,56 +12,8 @@
         <option value="$PROJECT_DIR$/analyzer/analyzer-bybit/pom.xml" />
         <option value="$PROJECT_DIR$/core/test-trader/test-trader-bybit/pom.xml" />
         <option value="$PROJECT_DIR$/core/trader/trader-bybit/pom.xml" />
+        <option value="$PROJECT_DIR$/exchange-connector/exchangr-connector-common/pom.xml" />
       </set>
-    </option>
-  </component>
-  <component name="NullableNotNullManager">
-    <option name="myDefaultNullable" value="org.jetbrains.annotations.Nullable" />
-    <option name="myDefaultNotNull" value="org.jetbrains.annotations.NotNull" />
-    <option name="myOrdered" value="false" />
-    <option name="myNullables">
-      <value>
-        <list size="16">
-          <item index="0" class="java.lang.String" itemvalue="org.jspecify.annotations.Nullable" />
-          <item index="1" class="java.lang.String" itemvalue="org.jetbrains.annotations.Nullable" />
-          <item index="2" class="java.lang.String" itemvalue="android.annotation.Nullable" />
-          <item index="3" class="java.lang.String" itemvalue="androidx.annotation.Nullable" />
-          <item index="4" class="java.lang.String" itemvalue="com.android.annotations.Nullable" />
-          <item index="5" class="java.lang.String" itemvalue="android.support.annotation.Nullable" />
-          <item index="6" class="java.lang.String" itemvalue="androidx.annotation.RecentlyNullable" />
-          <item index="7" class="java.lang.String" itemvalue="org.checkerframework.checker.nullness.qual.Nullable" />
-          <item index="8" class="java.lang.String" itemvalue="org.checkerframework.checker.nullness.compatqual.NullableDecl" />
-          <item index="9" class="java.lang.String" itemvalue="org.checkerframework.checker.nullness.compatqual.NullableType" />
-          <item index="10" class="java.lang.String" itemvalue="org.eclipse.jdt.annotation.Nullable" />
-          <item index="11" class="java.lang.String" itemvalue="edu.umd.cs.findbugs.annotations.Nullable" />
-          <item index="12" class="java.lang.String" itemvalue="jakarta.annotation.Nullable" />
-          <item index="13" class="java.lang.String" itemvalue="javax.annotation.Nullable" />
-          <item index="14" class="java.lang.String" itemvalue="javax.annotation.CheckForNull" />
-          <item index="15" class="java.lang.String" itemvalue="org.springframework.lang.Nullable" />
-        </list>
-      </value>
-    </option>
-    <option name="myNotNulls">
-      <value>
-        <list size="16">
-          <item index="0" class="java.lang.String" itemvalue="org.jspecify.annotations.NonNull" />
-          <item index="1" class="java.lang.String" itemvalue="org.jetbrains.annotations.NotNull" />
-          <item index="2" class="java.lang.String" itemvalue="android.annotation.NonNull" />
-          <item index="3" class="java.lang.String" itemvalue="androidx.annotation.NonNull" />
-          <item index="4" class="java.lang.String" itemvalue="com.android.annotations.NonNull" />
-          <item index="5" class="java.lang.String" itemvalue="android.support.annotation.NonNull" />
-          <item index="6" class="java.lang.String" itemvalue="androidx.annotation.RecentlyNonNull" />
-          <item index="7" class="java.lang.String" itemvalue="org.checkerframework.checker.nullness.qual.NonNull" />
-          <item index="8" class="java.lang.String" itemvalue="org.checkerframework.checker.nullness.compatqual.NonNullDecl" />
-          <item index="9" class="java.lang.String" itemvalue="org.checkerframework.checker.nullness.compatqual.NonNullType" />
-          <item index="10" class="java.lang.String" itemvalue="org.eclipse.jdt.annotation.NonNull" />
-          <item index="11" class="java.lang.String" itemvalue="edu.umd.cs.findbugs.annotations.NonNull" />
-          <item index="12" class="java.lang.String" itemvalue="jakarta.annotation.Nonnull" />
-          <item index="13" class="java.lang.String" itemvalue="javax.annotation.Nonnull" />
-          <item index="14" class="java.lang.String" itemvalue="lombok.NonNull" />
-          <item index="15" class="java.lang.String" itemvalue="org.springframework.lang.NonNull" />
-        </list>
-      </value>
     </option>
   </component>
   <component name="ProjectRootManager" version="2" project-jdk-name="openjdk-25" project-jdk-type="JavaSDK">
