--- common/src/main/java/io/github/ijlijapol/data/market/model/request/Symbol.java
@@ -1,5 +1,0 @@
-package io.github.ijlijapol.data.market.model.request;
-
-public enum Symbol {
-    BTCUSDT
-}
