--- common/src/main/java/io/github/ijlijapol/data/market/model/request/TimeFrame.java
@@ -1,5 +1,0 @@
-package io.github.ijlijapol.data.market.model.request;
-
-public enum TimeFrame {
-    FIFTEEN_MINUTES
-}
