+++ common/src/main/java/io/github/ijlijapol/model/Symbol.java
@@ -1,0 +1,5 @@
+package io.github.ijlijapol.model;
+
+public enum Symbol {
+    BTCUSDT
+}
