+++ common/src/main/java/io/github/ijlijapol/model/request/MarketDataForPeriodBetween.java
@@ -1,4 +1,4 @@
-package io.github.ijlijapol.data.market.model.request;
+package io.github.ijlijapol.model.request;
 
 import lombok.*;
 import lombok.experimental.SuperBuilder;
