+++ common/src/main/java/io/github/ijlijapol/model/request/TimeFrame.java
@@ -1,0 +1,5 @@
+package io.github.ijlijapol.model.request;
+
+public enum TimeFrame {
+    FIFTEEN_MINUTES
+}
