+++ market-data/market-data-bybit/src/main/java/io/github/ijlijapol/ByBitLoaderMarketDataImpl.java
@@ -5,13 +5,13 @@ import com.bybit.api.client.domain.market.MarketInterval;
 import com.bybit.api.client.domain.market.response.kline.MarketKlineEntry;
 import com.bybit.api.client.restApi.BybitApiMarketRestClient;
 import com.bybit.api.client.service.BybitApiClientFactory;
-import io.github.ijlijapol.data.market.LoaderMarketData;
-import io.github.ijlijapol.data.market.model.request.LastTime;
-import io.github.ijlijapol.data.market.model.request.MarketDataForPeriodBetween;
-import io.github.ijlijapol.data.market.model.request.MarketDataRequest;
-import io.github.ijlijapol.data.market.model.request.RecentMarketData;
-import io.github.ijlijapol.data.market.model.responce.CandleDTO;
-import io.github.ijlijapol.data.market.model.responce.CandlesDTO;
+import io.github.ijlijapol.contract.LoaderMarketData;
+import io.github.ijlijapol.model.request.LastTime;
+import io.github.ijlijapol.model.request.MarketDataForPeriodBetween;
+import io.github.ijlijapol.model.request.MarketDataRequest;
+import io.github.ijlijapol.model.request.RecentMarketData;
+import io.github.ijlijapol.model.responce.CandleDTO;
+import io.github.ijlijapol.model.responce.CandlesDTO;
 import io.github.ijlijapol.exception.ByBitException;
 import io.github.ijlijapol.exception.UncorrectedRequestByBit;
 import io.github.ijlijapol.mapper.MapperByBitData;
