+++ market-data/market-data-bybit/src/test/java/io/github/ijlijapol/ByBitLoaderMarketDataImplTest.java
@@ -4,9 +4,10 @@ import com.bybit.api.client.domain.CategoryType;
 import com.bybit.api.client.domain.market.MarketInterval;
 import com.bybit.api.client.domain.market.response.kline.MarketKlineEntry;
 import com.bybit.api.client.restApi.BybitApiMarketRestClient;
-import io.github.ijlijapol.data.market.model.request.*;
-import io.github.ijlijapol.data.market.model.responce.CandleDTO;
-import io.github.ijlijapol.data.market.model.responce.CandlesDTO;
+import io.github.ijlijapol.model.Symbol;
+import io.github.ijlijapol.model.request.*;
+import io.github.ijlijapol.model.responce.CandleDTO;
+import io.github.ijlijapol.model.responce.CandlesDTO;
 import io.github.ijlijapol.mapper.MapperByBitData;
 import io.github.ijlijapol.mapper.MapperTimeFrame;
 import org.junit.jupiter.api.BeforeEach;
